added comments and removed redundant code, also added debugging statements
