--- catalyst/src/App.css
@@ -1,70 +1,0 @@
-body{
-   background-color: #343b49;
-   color: rgba(228, 228, 228, 0.986);
-}
-
-.Homepage{
-  display: flex;
-}
-
-.WelcomeHeader{
-  display: flex;
-  height: 40vmin;
-  pointer-events: none;
-}
-
-.Questionnaire{
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  font-size: x-large;
-}
-
-
-.Welcome{
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-  padding-left: 3em;
-  padding-right: 3em;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-.ButtonContainer{
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  margin-top: 5em;
-  position: relative;
-  margin-bottom: 50vh;
-}
-.SubmitButton{
-  background-color:  rgb(162, 162, 247);
-  width: 150px;
-  height: 75px;
-}
-
-.ButtonText{
-  font-size: 2em;
-}
-
-.Form{
-  display: flex;
-  flex-direction: column;
-}
-
-select{
-  background-color: rgb(162, 162, 247);
-}
-
-.NameForm{
-  background-color: rgb(162, 162, 247);
-}
