+++ catalyst/src/Main.css
@@ -1,0 +1,31 @@
+body{
+   background-color: #343b49;
+   color: rgba(228, 228, 228, 0.986);
+}
+
+
+.ButtonContainer{
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  margin-top: 5em;
+  position: relative;
+  margin-bottom: 50vh;
+}
+
+.ButtonText{
+  font-size: 2em;
+}
+
+.Form{
+  display: flex;
+  flex-direction: column;
+}
+
+select{
+  background-color: rgb(162, 162, 247);
+}
+
+.NameForm{
+  background-color: rgb(162, 162, 247);
+}
