+++ catalyst/src/server.js
@@ -4,7 +4,6 @@ const mongoose = require('mongoose');
 const authRoutes = require('../routes/authRoutes');
 require('dotenv').config();
 
-
 const DB_URI = process.env.DB_URI;
 
 // Create Express application
@@ -12,7 +11,6 @@ const app = express();
 
 // Middleware setup
 app.use(express.json());
-
 
 // Database connection
 mongoose.connect(DB_URI, {
@@ -22,38 +20,11 @@ mongoose.connect(DB_URI, {
     .then(() => console.log('Connected to MongoDB'))
     .catch((err) => console.error('Failed to connect to MongoDB', err));
 
-// Additional routes and middleware can be added here...
+// Use authRoutes middleware
+app.use('/api/user', authRoutes);
 
 // Start the server
 const PORT = process.env.PORT || 3000;
-app.listen(PORT, () => {
+app.listen(PORT, '0.0.0.0', () => {
     console.log(`Server is running on port ${PORT}`);
 });
-
-app.post('/register', async (req, res) => {
-    const { username, email, password } = req.body;
-
-    try {
-        const user = new User({ username, email, password });
-        await user.save();
-        res.status(201).json({ message: 'User registered successfully' });
-    } catch (error) {
-        res.status(500).json({ error: 'Internal server error' });
-    }
-});
-
-// Middleware
-app.use('/api/user', authRoutes);
-
-// server.js or routes/auth.js
-
-const router = express.Router();
-
-// Define route for user registration
-router.post('/register', (req, res) => {
-    // Handle user registration logic here
-});
-
-module.exports = router;
-
-
